[web] refactor: components folders
